Implemented profilePage.html
Added profilePage.html, this page displays a user's profile.  Includes a profile picture, name, table showing profile information, and a table showing the user's listings.  Problems: padding on tables is a little wonky, still working on a solution but for now I think it's fine. Also, for some reason I can't get a padding on the bottom of the page, but again that's just an asthetics issue.  Resizing page is okay... could use some work.